refactor: updating packages
