+++ packages/cli/package.json
@@ -1,35 +1,35 @@
 {
   "name": "@bluebase/cli",
   "description": "CLI tool to make life with BlueBase easier!",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "author": "BlueEast Team <code@blueeast.com>",
   "bin": {
     "bluebase": "./bin/run"
   },
   "bugs": "https://github.com/BlueBaseJS/cli/issues",
   "dependencies": {
-    "@oclif/command": "^1",
-    "@oclif/config": "^1",
+    "@oclif/command": "^1.5.8",
+    "@oclif/config": "^1.10.4",
     "@oclif/plugin-help": "^2",
     "@oclif/plugin-plugins": "^1.5.4",
     "inquirer": "^6.2.0",
     "tslib": "^1"
   },
   "devDependencies": {
-    "@oclif/dev-cli": "^1",
+    "@oclif/dev-cli": "^1.21.0",
     "@oclif/test": "^1",
-    "@oclif/tslint": "^1",
+    "@oclif/tslint": "^3.1.1",
     "@types/chai": "^4",
     "@types/inquirer": "^0.0.43",
     "@types/mocha": "^5",
-    "@types/node": "^10",
+    "@types/node": "^10.12.18",
     "chai": "^4",
     "globby": "^8",
     "mocha": "^5",
-    "nyc": "^12",
-    "ts-node": "^6",
-    "typescript": "^3.2.1",
-		"tslint": "^5.11.0"
+    "nyc": "^13.1.0",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.12.0",
+    "typescript": "^3.2.2"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -54,7 +54,7 @@
       "@oclif/plugin-plugins"
     ]
   },
-  "repository": "BlueBaseJS/cli",
+  "repository": "https://github.com/BlueBaseJS/cli",
   "scripts": {
     "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
     "posttest": "tsc -p test --noEmit && tslint -p test -t stylish",
