+++ packages/core/README.md
