+++ packages/core/package.json
@@ -1,18 +1,18 @@
 {
   "name": "@bluebase/cli-core",
   "description": "A plugin for BlueBase CLI",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "author": "BlueEast Team <code@blueeast.com>",
   "main": "./lib",
   "sideEffects": false,
   "bugs": "https://github.com/BlueBaseJS/cli/issues",
   "dependencies": {
-    "@oclif/command": "^1",
-    "@oclif/config": "^1",
-    "deepmerge": "^2.1.1",
+    "@oclif/command": "^1.5.8",
+    "@oclif/config": "^1.10.4",
+    "deepmerge": "^3.0.0",
     "detect-installed": "^2.0.4",
-    "dotenv": "^6.0.0",
-    "file-regex": "^2.1.1",
+    "dotenv": "^6.2.0",
+    "file-regex": "^3.0.1",
     "inquirer": "^6.2.0",
     "lodash.isnil": "^4.0.0",
     "lodash.template": "^4.4.0",
@@ -23,31 +23,32 @@
     "yarn-install": "^1.0.0"
   },
   "devDependencies": {
-    "@blueeast/tslint-config-blueeast": "^0.5.6",
-    "@oclif/dev-cli": "^1",
+    "@blueeast/tslint-config-blueeast": "^0.7.0",
+    "@oclif/dev-cli": "^1.21.0",
     "@oclif/plugin-help": "^2",
     "@oclif/test": "^1",
-    "@oclif/tslint": "^1",
+    "@oclif/tslint": "^3.1.1",
     "@types/app-root-dir": "^0.1.0",
     "@types/chai": "^4",
     "@types/deepmerge": "^2.1.0",
-    "@types/dotenv": "^4.0.3",
+    "@types/dotenv": "^6.1.0",
     "@types/inquirer": "^0.0.43",
     "@types/lodash.isnil": "^4.0.4",
     "@types/lodash.template": "^4.4.4",
     "@types/mocha": "^5",
-    "@types/node": "^10.7.0",
+    "@types/node": "^10.12.18",
     "@types/node-notifier": "^0.0.28",
-    "@types/react": "^16.4.10",
-    "@types/shelljs": "^0.8.0",
+    "@types/react": "^16.7.18",
+    "@types/shelljs": "^0.8.1",
     "@types/winston": "^2.4.4",
     "chai": "^4",
+    "jest": "^23.6.0",
     "mocha": "^5",
     "npm-run-all": "^4.1.3",
-    "nyc": "^12",
-    "ts-node": "^6",
-		"tslint": "^5.11.0",
-    "typescript": "^3.2.1"
+    "nyc": "^13.1.0",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.12.0",
+    "typescript": "^3.2.2"
   },
   "peerDependencies": {
     "react": "^16.*"
@@ -74,7 +75,7 @@
       "@oclif/plugin-help"
     ]
   },
-  "repository": "BlueBaseJS/cli",
+  "repository": "https://github.com/BlueBaseJS/cli",
   "scripts": {
     "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
     "posttest": "tsc -p test --noEmit && tslint -p test -t stylish",
