+++ packages/core/src/FileManager/bluebase/HookRegistry.ts
