+++ packages/core/src/FileManager/bluebase/Registry.ts
