--- packages/core/test/commands/hello.test.ts
@@ -1,17 +1,0 @@
-import {expect, test} from '@oclif/test'
-
-describe('hello', () => {
-  test
-  .stdout()
-  .command(['hello'])
-  .it('runs hello', ctx => {
-    expect(ctx.stdout).to.contain('hello world')
-  })
-
-  test
-  .stdout()
-  .command(['hello', '--name', 'jeff'])
-  .it('runs hello --name jeff', ctx => {
-    expect(ctx.stdout).to.contain('hello jeff')
-  })
-})
