--- packages/core/test/helpers/init.js
@@ -1,2 +1,0 @@
-const path = require('path')
-process.env.TS_NODE_PROJECT = path.resolve('test/tsconfig.json')
