--- packages/core/test/mocha.opts
@@ -1,7 +1,0 @@
---require test/helpers/init.js
---require ts-node/register
---require source-map-support/register
---watch-extensions ts
---recursive
---reporter spec
---timeout 5000
