--- packages/core/test/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "declaration": false,
-  "extends": "../tsconfig",
-  "compilerOptions": {
-    "sourceMap": true
-  },
-  "include": [
-    "./**/*",
-    "../src/**/*"
-  ]
-}
