+++ packages/expo/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@bluebase/cli-expo",
   "description": "An Expo plugin for BlueBaseJS CLI",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "main": "./lib",
   "author": "BlueEast Team <code@blueeast.com>",
   "bugs": "https://github.com/BlueBaseJS/cli/issues",
@@ -65,7 +65,7 @@
       "@oclif/plugin-help"
     ]
   },
-  "repository": "BlueBaseJS/cli",
+  "repository": "https://github.com/BlueBaseJS/cli",
   "scripts": {
     "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
     "posttest": "tsc -p test --noEmit && tslint -p test -t stylish",
