+++ packages/storybook/templates/storybook/sample/DummyComponent.stories.tsx
@@ -1,3 +1,5 @@
+declare var module: any;
+
 import { BlueBase, BlueBaseConsumer } from '@bluebase/core';
 import React from 'react';
 import storiesOf from '@bluebase/storybook-addon';
