+++ packages/web/package.json
@@ -1,15 +1,15 @@
 {
   "name": "@bluebase/cli-web",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "author": "BlueEast Team <code@blueeast.com>",
   "bugs": "https://github.com/BlueBaseJS/cli/issues",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "module": "lib/index.js",
   "dependencies": {
-    "@bluebase/cli-core": "^2.0.0-beta.12",
-    "@oclif/command": "^1",
-    "@oclif/config": "^1",
+    "@bluebase/cli-core": "^0.0.1",
+    "@oclif/command": "^1.5.8",
+    "@oclif/config": "^1.10.4",
     "@types/html-webpack-plugin": "^3.2.0",
     "@types/webpack-serve": "^2.0.0",
     "assets-webpack-plugin": "^3.8.4",
@@ -21,7 +21,7 @@
     "compression": "^1.7.3",
     "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^1.0.0",
-    "deepmerge": "^2.1.1",
+    "deepmerge": "^3.0.0",
     "express": "^4.16.3",
     "file-loader": "^1.1.11",
     "fork-ts-checker-webpack-plugin": "^0.4.4",
@@ -35,11 +35,11 @@
     "normalize.css": "^8.0.0",
     "offline-plugin": "^5.0.5",
     "optimize-css-assets-webpack-plugin": "^5.0.0",
-    "react-art": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "react-art": "^16.7.0",
+    "react-dom": "^16.7.0",
     "react-helmet": "^5.2.0",
-    "react-hot-loader": "^4.3.4",
-    "react-native-web": "^0.8.9",
+    "react-hot-loader": "^4.6.3",
+    "react-native-web": "^0.9.10",
     "rmfr": "^2.0.0",
     "speed-measure-webpack-plugin": "^1.2.2",
     "style-loader": "^0.21.0",
@@ -61,71 +61,68 @@
     "write-json-webpack-plugin": "^1.0.2"
   },
   "devDependencies": {
-    "@bluebase/core": "^3.13.7",
-    "@blueeast/bluerain-platform-react-native": "^1.1.3",
-    "@blueeast/bluerain-plugin-react-router": "^1.1.5",
-    "@blueeast/bluerain-plugin-redux": "^0.6.6",
-    "@blueeast/tslint-config-blueeast": "^0.5.6",
-    "@oclif/dev-cli": "^1",
+    "@bluebase/core": "^4.0.0-aplha.4",
+    "@blueeast/tslint-config-blueeast": "^0.7.0",
+    "@oclif/dev-cli": "^1.21.0",
     "@oclif/plugin-help": "^2",
     "@oclif/test": "^1",
-    "@oclif/tslint": "^1",
+    "@oclif/tslint": "^3.1.1",
     "@types/assets-webpack-plugin": "^3.5.1",
     "@types/chai": "^4",
     "@types/chokidar": "^1.7.5",
     "@types/clean-webpack-plugin": "^0.1.2",
     "@types/compression": "^0.0.36",
     "@types/copy-webpack-plugin": "^4.4.2",
-    "@types/debug": "^0.0.30",
+    "@types/debug": "^0.0.31",
     "@types/deepmerge": "^2.1.0",
     "@types/express": "^4.16.0",
     "@types/fork-ts-checker-webpack-plugin": "^0.4.1",
-    "@types/helmet": "^0.0.38",
+    "@types/helmet": "^0.0.42",
     "@types/hpp": "^0.2.1",
     "@types/jest": "^23.1.6",
     "@types/md5": "^2.1.32",
     "@types/mini-css-extract-plugin": "^0.2.0",
     "@types/mocha": "^5",
-    "@types/node": "^10.5.2",
+    "@types/node": "^10.12.18",
     "@types/node-notifier": "^0.0.28",
     "@types/optimize-css-assets-webpack-plugin": "^1.3.3",
-    "@types/react": "^16.4.7",
+    "@types/react": "^16.7.18",
     "@types/react-dom": "^16.0.6",
     "@types/react-helmet": "^5.0.7",
     "@types/rmfr": "^2.0.0",
     "@types/serialize-javascript": "^1.3.2",
     "@types/uglifyjs-webpack-plugin": "^1.1.0",
     "@types/uuid": "^3.4.3",
-    "@types/webpack": "^4.4.9",
+    "@types/webpack": "^4.4.22",
     "@types/webpack-dev-middleware": "^2.0.2",
     "@types/webpack-hot-middleware": "^2.16.4",
     "@types/webpack-merge": "^4.1.3",
     "@types/webpack-node-externals": "^1.6.3",
     "chai": "^4",
     "codecov": "^3.0.2",
-    "commitizen": "^2.10.1",
+    "commitizen": "^3.0.5",
     "cz-conventional-changelog": "^2.1.0",
-    "gh-pages": "^1.2.0",
+    "gh-pages": "^2.0.1",
     "globby": "^8",
     "jest": "^23.4.1",
     "mocha": "^5",
     "npm-run-all": "^4.1.3",
-    "nyc": "^12",
-    "opn-cli": "^3.1.0",
+    "nyc": "^13.1.0",
+    "opn-cli": "^4.0.0",
     "prettier": "^1.13.4",
-    "react": "^16.4.2",
+    "react": "^16.7.0",
     "rimraf": "^2.6.2",
     "standard-version": "^4.4.0",
     "trash-cli": "^1.4.0",
     "ts-node": "^7.0.0",
-		"tslint": "^5.11.0",
+    "tslint": "^5.12.0",
     "tslint-config-prettier": "^1.13.0",
-    "tslint-immutable": "^4.5.4",
-    "typedoc": "^0.11.1",
-    "typescript": "^3.2.1"
+    "tslint-immutable": "^5.0.1",
+    "typedoc": "^0.13.0",
+    "typescript": "^3.2.2"
   },
   "peerDependencies": {
-    "@bluebase/core": "^3.*",
+    "@bluebase/core": "^4.*",
     "react": "^16.*",
     "react-dom": "^16.*"
   },
@@ -156,7 +153,7 @@
       "@oclif/plugin-help"
     ]
   },
-  "repository": "BlueBaseJS/cli",
+  "repository": "https://github.com/BlueBaseJS/cli",
   "scripts": {
     "postpack": "rm -f oclif.manifest.json npm-shrinkwrap.json",
     "posttest": "tsc -p test --noEmit && tslint -p test -t stylish",
