+++ packages/web/src/client/ReactHotLoaderPlugin/HOC/index.ts
@@ -1,12 +1,12 @@
-import { Plugin } from '@bluebase/core';
-import { hot } from 'react-hot-loader';
+// import { Plugin } from '@bluebase/core';
+// import { hot } from 'react-hot-loader';
 
-export default class ReactHotLoaderPlugin extends Plugin {
-​
-	static pluginName = 'ReactHotLoaderPlugin';
-	static slug = 'react-hot-loader';
-​
-	static initialize(_config: any, BR: any) {
-		BR.Components.addHocs('SystemApp', hot(module));
-	}
-}
+// export default class ReactHotLoaderPlugin extends Plugin {
+// ​
+// 	static pluginName = 'ReactHotLoaderPlugin';
+// 	static slug = 'react-hot-loader';
+// ​
+// 	static initialize(_config: any, BR: any) {
+// 		BR.Components.addHocs('SystemApp', hot(module));
+// 	}
+// }
